Move some tests over in individual files
