+++ tests/IrcMessageTest.php
@@ -54,256 +54,6 @@ use WildPHP\Messages\WhoWas;
 
 class IrcMessageTest extends TestCase
 {
-	public function testAccountCreate()
-	{
-		$account = new Account('ircAccount');
-
-		$this->assertEquals('ircAccount', $account->getAccountName());
-
-		$prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-		$this->expectException(\InvalidArgumentException::class);
-		Account::fromIncomingMessage($incomingIrcMessage);
-	}
-
-	public function testAccountReceive()
-	{
-        $prefix = 'nickname!username@hostname';
-        $verb = 'ACCOUNT';
-        $args = ['ircAccount'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$account = Account::fromIncomingMessage($incoming);
-
-		$userPrefix = new Prefix('nickname', 'username', 'hostname');
-		$this->assertEquals($userPrefix, $account->getPrefix());
-		$this->assertEquals('ircAccount', $account->getAccountName());
-	}
-
-	public function testAuthenticateCreate()
-	{
-		$authenticate = new Authenticate('+');
-
-		$this->assertEquals('+', $authenticate->getResponse());
-
-		$expected = 'Authenticate +' . "\r\n";
-		$this->assertEquals($expected, $authenticate->__toString());
-	}
-
-	public function testAuthenticateReceive()
-	{
-        $prefix = '';
-        $verb = 'AUTHENTICATE';
-        $args = ['+'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$authenticate = Authenticate::fromIncomingMessage($incoming);
-
-		$this->assertEquals('+', $authenticate->getResponse());
-
-		$prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-		$this->expectException(\InvalidArgumentException::class);
-		Authenticate::fromIncomingMessage($incomingIrcMessage);
-	}
-
-	public function testAwayCreate()
-	{
-		$away = new Away('A sample message');
-
-		$this->assertEquals('A sample message', $away->getMessage());
-
-		$expected = 'Away :A sample message' . "\r\n";
-		$this->assertEquals($expected, $away->__toString());
-	}
-
-	public function testAwayReceive()
-	{
-        $prefix = 'nickname!username@hostname';
-        $verb = 'AWAY';
-        $args = ['A sample message'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$away = Away::fromIncomingMessage($incoming);
-
-		$this->assertEquals('nickname', $away->getNickname());
-		$this->assertEquals('A sample message', $away->getMessage());
-
-		$prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-		$this->expectException(\InvalidArgumentException::class);
-		Away::fromIncomingMessage($incomingIrcMessage);
-	}
-
-	public function testCapCreate()
-    {
-        $cap = new Cap('REQ', ['cap1', 'cap2']);
-
-        $this->assertEquals('REQ', $cap->getCommand());
-        $this->assertEquals(['cap1', 'cap2'], $cap->getCapabilities());
-        $this->assertTrue($cap->isFinalMessage());
-
-        $expected = 'CAP REQ :cap1 cap2' . "\r\n";
-        $this->assertEquals($expected, $cap->__toString());
-
-        $cap = new Cap('REQ', ['cap1', 'cap2'], false);
-        $this->assertFalse($cap->isFinalMessage());
-    }
-
-	public function testCapCreateInvalidSubcommand()
-	{
-	    new Cap('LS 302');
-
-		$this->expectException(\InvalidArgumentException::class);
-		
-		new Cap('INVALID');
-    }
-
-    public function testCapReceive()
-    {
-        $prefix = 'server';
-        $verb = 'CAP';
-        $args = ['*', 'LS', 'cap1 cap2'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-        $cap = Cap::fromIncomingMessage($incoming);
-
-        $this->assertEquals('LS', $cap->getCommand());
-        $this->assertEquals(['cap1', 'cap2'], $cap->getCapabilities());
-        $this->assertEquals('*', $cap->getClientIdentifier());
-        $this->assertTrue($cap->isFinalMessage());
-
-        $prefix = 'server';
-        $verb = 'CAP';
-        $args = ['*', 'LS', '*', 'cap1 cap2'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-        $cap = Cap::fromIncomingMessage($incoming);
-
-        $this->assertEquals('LS', $cap->getCommand());
-        $this->assertEquals(['cap1', 'cap2'], $cap->getCapabilities());
-        $this->assertEquals('*', $cap->getClientIdentifier());
-        $this->assertFalse($cap->isFinalMessage());
-
-	    $prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-	    $this->expectException(\InvalidArgumentException::class);
-	    Cap::fromIncomingMessage($incomingIrcMessage);
-    }
-
-	public function testErrorReceive()
-	{
-        $prefix = '';
-        $verb = 'ERROR';
-        $args = ['A sample message'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$error = Error::fromIncomingMessage($incoming);
-
-		$this->assertEquals('A sample message', $error->getMessage());
-
-		$prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-		$this->expectException(\InvalidArgumentException::class);
-		Error::fromIncomingMessage($incomingIrcMessage);
-	}
-
-	public function testJoinCreate()
-	{
-		$join = new Join(['#channel1', '#channel2'], ['key1', 'key2']);
-
-		$this->assertEquals(['#channel1', '#channel2'], $join->getChannels());
-		$this->assertEquals(['key1', 'key2'], $join->getKeys());
-
-		$expected = 'Join #channel1,#channel2 key1,key2' . "\r\n";
-		$this->assertEquals($expected, $join->__toString());
-
-		$join = new Join('#channel1', 'key1');
-
-		$this->assertEquals(['#channel1'], $join->getChannels());
-		$this->assertEquals(['key1'], $join->getKeys());
-	}
-
-	public function testJoinCreateKeyMismatch()
-	{
-		$this->expectException(InvalidArgumentException::class);
-
-		new Join(['#channel1', '#channel2'], ['key1']);
-	}
-
-	public function testJoinReceiveExtended()
-	{
-		$prefix = 'nickname!username@hostname';
-        $verb = 'JOIN';
-        $args = ['#channel', 'ircAccountName', 'realname'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$join = Join::fromIncomingMessage($incoming);
-
-		$this->assertEquals('nickname', $join->getNickname());
-		$this->assertEquals(['#channel'], $join->getChannels());
-		$this->assertEquals('ircAccountName', $join->getIrcAccount());
-		$this->assertEquals('realname', $join->getRealname());
-	}
-
-	public function testJoinReceiveRegular()
-	{
-        $prefix = 'nickname!username@hostname';
-        $verb = 'JOIN';
-        $args = ['#channel'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$join = Join::fromIncomingMessage($incoming);
-
-		$this->assertEquals('nickname', $join->getNickname());
-		$this->assertEquals(['#channel'], $join->getChannels());
-		$this->assertEquals('', $join->getIrcAccount());
-		$this->assertEquals('', $join->getRealname());
-
-		$prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-		$this->expectException(\InvalidArgumentException::class);
-		Join::fromIncomingMessage($incomingIrcMessage);
-	}
-
-	public function testKickCreate()
-	{
-		$kick = new Kick('#channel', 'nickname', 'Bleep you!');
-
-		$this->assertEquals('#channel', $kick->getChannel());
-		$this->assertEquals('nickname', $kick->getTarget());
-		$this->assertEquals('Bleep you!', $kick->getMessage());
-
-		$expected = 'KICK #channel nickname :Bleep you!' . "\r\n";
-		$this->assertEquals($expected, $kick->__toString());
-	}
-
-	public function testKickReceive()
-	{
-        $prefix = 'nickname!username@hostname';
-        $verb = 'KICK';
-        $args = ['#somechannel', 'othernickname', 'You deserved it!'];
-        $incoming = new IrcMessage($prefix, $verb, $args);
-		$kick = Kick::fromIncomingMessage($incoming);
-
-		$userPrefix = new Prefix('nickname', 'username', 'hostname');
-		$this->assertEquals($userPrefix, $kick->getPrefix());
-		$this->assertEquals('nickname', $kick->getNickname());
-		$this->assertEquals('othernickname', $kick->getTarget());
-		$this->assertEquals('#somechannel', $kick->getChannel());
-		$this->assertEquals('You deserved it!', $kick->getMessage());
-
-		$prefix = ':server';
-		$verb = 'TEEHEE';
-		$args = ['argument'];
-        $incomingIrcMessage = new IrcMessage($prefix, $verb, $args);
-		$this->expectException(\InvalidArgumentException::class);
-		Kick::fromIncomingMessage($incomingIrcMessage);
-	}
 
 	public function testModeCreate()
 	{
